# impl: avoid loading noop classes