Tested with the Benchmark allowed to load the java7/9 packages

```
BEFORE

Benchmark                                             Mode  Cnt     Score     Error        Units
ClassInitBenchmark.forName_init                         ss  400  8730.958 ± 142.833        us/op
ClassInitBenchmark.forName_init:·class.load             ss  400     1.166 ±   0.015  classes/sec
ClassInitBenchmark.forName_init:·class.load.norm        ss  400    72.105 ±   0.112   classes/op
ClassInitBenchmark.forName_init:·class.unload           ss  400       ≈ 0            classes/sec
ClassInitBenchmark.forName_init:·class.unload.norm      ss  400       ≈ 0             classes/op
ClassInitBenchmark.forName_noinit                       ss  400  3040.308 ±  42.065        us/op
ClassInitBenchmark.forName_noinit:·class.load           ss  400     0.605 ±   0.006  classes/sec
ClassInitBenchmark.forName_noinit:·class.load.norm      ss  400    57.845 ±   0.096   classes/op
ClassInitBenchmark.forName_noinit:·class.unload         ss  400       ≈ 0            classes/sec
ClassInitBenchmark.forName_noinit:·class.unload.norm    ss  400       ≈ 0             classes/op

AFTER

Benchmark                                             Mode  Cnt     Score     Error        Units
ClassInitBenchmark.forName_init                         ss  400  8411.036 ± 148.305        us/op
ClassInitBenchmark.forName_init:·class.load             ss  400     1.136 ±   0.016  classes/sec
ClassInitBenchmark.forName_init:·class.load.norm        ss  400    71.040 ±   0.113   classes/op
ClassInitBenchmark.forName_init:·class.unload           ss  400       ≈ 0            classes/sec
ClassInitBenchmark.forName_init:·class.unload.norm      ss  400       ≈ 0             classes/op
ClassInitBenchmark.forName_noinit                       ss  400  3077.689 ±  52.719        us/op
ClassInitBenchmark.forName_noinit:·class.load           ss  400     0.612 ±   0.007  classes/sec
ClassInitBenchmark.forName_noinit:·class.load.norm      ss  400    57.815 ±   0.093   classes/op
ClassInitBenchmark.forName_noinit:·class.unload         ss  400       ≈ 0            classes/sec
ClassInitBenchmark.forName_noinit:·class.unload.norm    ss  400       ≈ 0             classes/op
```